nfs-ganesha: rip out EXPORT handling
Now that I've had some opportunity to work with ceph-mgr and the
orchestrator, I think we'd be best served by simply having rook
spin up the ganesha containers for us and leave the export management
to other layers.

Remove all of the export handing code in the nfs-ganesha CRD and just
have it generate a mostly-stock config for all of the daemons.

Signed-off-by: Jeff Layton <jlayton@redhat.com>